Courses microservice inserts successfully data into the database

+++ microservice-course/src/main/java/com/microservice/course/persistence/ICourseRepository.java
@@ -4,8 +4,6 @@ import com.microservice.course.entities.Course;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
-
 @Repository
-public interface ICourseRepository extends CrudRepository<Course, BigDecimal> {
+public interface ICourseRepository extends CrudRepository<Course, Long> {
 }
